requirements.txt
torch
torchvision
easydict
pandas
numpy
scipy
matplotlib
Pillow
opencv-python
scikit-image